When a supplier modified a taxpayer, then the taxpayer change his status to 'changes pending' Add workday_id unique constraint and msg error in case of workday id repeated Moving constants to 'constansts' directory
