+++ supplier_app/constants/custom_messages.py
@@ -17,3 +17,4 @@ TAXPAYER_FORM_INVALID_MESSAGE = _("Your form is invalid")
 TAXPAYER_NOT_EXISTS_MESSAGE = _("The taxpayer you try to access/modify doesnt exists")
 TAXPAYER_REQUEST_CHANGE_MESSAGE = _("Requested changes successfully")
 TAXPAYER_WITHOUT_WORKDAY_ID_MESSAGE = _("You did not set workday id")
+TAXPAYER_WORKDAY_UNIQUE_ERROR = _("There is other taxpayer with that workday id")
