+++ supplier_app/exceptions/taxpayer_exceptions.py
@@ -1,2 +1,6 @@
 class NoWorkdayIDException(Exception):
     pass
+
+
+class TaxpayerUniqueWorkdayId(Exception):
+    pass
