+++ templates/AP_app/approve-taxpayer.html
@@ -1,7 +1,7 @@
 <form method="POST" action="{% url 'handle-taxpayer-status' taxpayer_id=taxpayer.id %}">
     {% csrf_token %}
     <div class="input-group">
-        <input type="number" required class="form-control" name="workday_id" placeholder="Workday ID">
+        <input type="number" class="form-control" name="workday_id" placeholder="Workday ID">
 	<input type="hidden" name="action" value="{{approve}}">                
         <input class='btn btn-success' type="submit" value="Approve" id="approve"/>
     </div>
